Break three simple import cycles by moving definitions between files
